Fix jupyter launching, and install adaptive packages
Jupyter labs had to be called and the webpage can be called
by using the port and ip directly. However, it might not be intuitice
for a student to enter the token number generated from the command lines
For this, the Jupyter Object automatically filters the URL with the token
number and accordingly opens in the WebView

WebKit activity and webview was migrated to browse activiy, for it provides
a lot of options and easier sugar3 integration.